Quizzcard and user quizz page
